requirements/base.txt
SQLAlchemy>=1.4.22
babel
psycopg2-binary
databases[asyncpg]
python-dotenv
pydantic
fastapi-utils
requests
alembic
pyfa-converter
passlib
uvicorn
bcrypt
jinja2
numpy
pandas 
scikit-learn
scikit-surprise